==> rmd/3_Simulation_SparseDOSSA.Rmd <==
---
title: "Simulation task 3 (SparseDOSSA model)"
author: "Siyuan Ma"
output: html_document
---

```{r set up}
knitr::opts_chunk$set(echo=FALSE)
library(magrittr)
```

```{r batchtools}
# Registry
# batchtools::makeRegistry(file.dir = "r_batchtools_reg/sim/SparseDOSSA",
#                          package = c("magrittr"))
batchtools::loadRegistry(file.dir = "r_batchtools_reg/sim/SparseDOSSA",
                         writeable = TRUE)
rm(list = ls())
dir_output <- "results/Simulation/SparseDOSSA"
dir.create(dir_output, showWarnings = TRUE, recursive = TRUE)
```

```{r true parameters}
# Prepare SparseDOSSA2 objects to simulate
Stool <- SparseDOSSA2:::Stool
fit_Stool <- Stool$EM_fit$fit
feature_full <- names(fit_Stool$pi0) %>% 
  stringr::str_replace_all("^.*\\|s\\_\\_", "") %>% 
  stringr::str_replace_all(stringr::fixed("_"), " ")
names(fit_Stool$pi0) <- 
  names(fit_Stool$mu) <- 
  names(fit_Stool$sigma) <- 
  rownames(fit_Stool$Sigma) <- 
  colnames(fit_Stool$Sigma) <- 
  rownames(fit_Stool$Omega) <- 
  colnames(fit_Stool$Omega) <- feature_full

feature_full <- 
  order(-(1 - fit_Stool$pi0) * fit_Stool$mu)[1:200] %>% 
  {feature_full[.]}
fit_Stool$pi0 <- fit_Stool$pi0[feature_full]
fit_Stool$mu <- fit_Stool$mu[feature_full]
fit_Stool$sigma <- fit_Stool$sigma[feature_full]
fit_Stool$Sigma <- fit_Stool$Sigma[feature_full, feature_full]
fit_Stool$Omega <- solve(fit_Stool$Sigma)
Stool$EM_fit$fit <- fit_Stool

# Features to be modified
tb_HMPII <- readr::read_csv("data/HMPII/cor.csv")
feature_up <- tb_HMPII %>% 
  dplyr::filter(`Coefficient (CD)` > 0) %>% 
  dplyr::arrange(-`Coefficient (CD)`) %>% 
  extract2("Feature")
feature_down <- tb_HMPII %>% 
  dplyr::filter(`Coefficient (CD)` < 0) %>% 
  dplyr::arrange(`Coefficient (CD)`) %>% 
  extract2("Feature") 
feature_up <- intersect(feature_up, feature_full)[seq(1, 10)]
feature_down <- intersect(feature_down, feature_full)[seq(1, 10)]
save(feature_full, file = paste0(dir_output, "/feature_full.RData"))
save(feature_up, file = paste0(dir_output, "/feature_up.RData"))
save(feature_down, file = paste0(dir_output, "/feature_down.RData"))
save(Stool, file = paste0(dir_output, "/Stool.RData"))
rm(list = c("fit_Stool", "IBD", "Vaginal", "tb_HMPII"))
```

```{r jobs}
tb_job <- tidyr::crossing(
  strength_disease = seq(0, 1.25, by = 0.25),
  strength_subject = 1,
  n_s = c(100, 30),
  n_t = 10,
  median_depth = c(1e4, 1e5),
  R = 2,
  rep = seq(1, 1000)
) %>% 
  dplyr::mutate(i_job = seq(1, dplyr::n()))
save(tb_job, file = paste0(dir_output, "/tb_job.RData"))
n_job <- nrow(tb_job)
n_job_each <- 100
rm(list = c("tb_job"))
```

```{r one job}
one_job <- function(i_job) {
  library(magrittr)
  load(paste0(dir_output, "/tb_job.RData"))
  load(paste0(dir_output, "/Stool.RData"))
  
  l_tb_pvals <- list()
  for(ii_job in seq((i_job - 1) * n_job_each + 1, i_job * n_job_each)) {
    set.seed(ii_job)
    i_tb_job <- tb_job[ii_job, ]
    
    #SparseDOSSA
    n_m <- length(feature_full)
    n_s <- i_tb_job$n_s
    n_t <- i_tb_job$n_t
    mat_metadata <- cbind(
      rep(rep(c(0, 1), each = n_s / 2), each = n_t),
      vapply(seq(1, 20),
             function(i) {
               rep(rnorm(n = n_s), each = n_t)
             },
             rep(0.0, length = n_s * n_t)))
    colnames(mat_metadata) <- c("disease",
                                paste0("subject_feature_", seq(1, 20)))
    df_metadata_spike <- 
      rbind(
        data.frame(metadata_datum = 1,
                   feature_spiked = feature_up,
                   effect_size = i_tb_job$strength_disease),
        data.frame(metadata_datum = 1,
                   feature_spiked = feature_down,
                   effect_size = -i_tb_job$strength_disease),
        data.frame(metadata_datum = seq(2, 21),
                   feature_spiked = 
                     sample(
                       feature_full, 
                       size = 20,
                       replace = FALSE),
                   effect_size = i_tb_job$strength_subject)
      ) %>% 
      tidyr::crossing(associated_property = c("prevalence", "abundance"))
    
    sim_SparseDOSSA2 <- 
      SparseDOSSA2::SparseDOSSA2(
        template = Stool,
        n_sample = nrow(mat_metadata), 
        new_features = FALSE,
        spike_metadata = df_metadata_spike,
        metadata_matrix = mat_metadata, 
        median_read_depth = i_tb_job$median_depth, 
        verbose = FALSE)
    
    mat_count_sim <- sim_SparseDOSSA2$simulated_data
    mat_norm <- apply(mat_count_sim, 2, function(x) x / sum(x))
    X_array_sim <- array(NA, dim = c(nrow(mat_count_sim),
                                     n_s, 
                                     n_t))
    for(j in seq(1, n_s))
      for(k in seq(1, n_t))
        X_array_sim[, j, k] <- mat_count_sim[, (j - 1) * n_t + k]
    
    
    # fit PCA
    fit_pca <- microTensor::pca(
      X = X_array_sim, R = i_tb_job$R)
    
    # fit ctf
    fit_ctf <- microTensor::ctf(
      X = X_array_sim,
      R = i_tb_job$R,
      control = list(maxit = 5000,
                     verbose = FALSE))
    # save(fit_ctf, file = paste0(dir_output, "/fit_ctf_", ii_job, ".RData"))
    
    # fit microTensor
    fit_microTensor <- microTensor::microTensor(
      X = X_array_sim,
      R = i_tb_job$R, 
      weighted = TRUE, ortho_m = FALSE, nn_t = FALSE,
      control = list(L_init = 2,
                     gamma = 2,
                     maxit = 5000,
                     verbose = FALSE))
    # save(fit_microTensor, 
    #      file = paste0(dir_output, "/fit_microTensor_", ii_job, ".RData"))
    
    # test on the first two axes
    tb_pvals_pca <- seq(1, i_tb_job$R) %>% 
      purrr::map_dfr(function(i_axis) {
        df_lme4 <- data.frame(
          outcome = fit_pca$s_full[, i_axis],
          exposure = rep(rep(c(0, 1), each = n_s / 2), n_t),
          subject = rep(seq(1, n_s), n_t))
        fit_lme4 <- lme4::lmer(
          outcome ~ exposure + (1|subject), 
          data = df_lme4)
        p <- lmerTest:::summary.lmerModLmerTest(fit_lme4)$coef[2, 5]
        tibble::tibble(p = p,
                       method = "PCA",
                       axis = i_axis)
      })
    
    tb_pvals <- seq(1, i_tb_job$R) %>% 
      purrr::map_dfr(function(i_axis) {
        tibble::tibble(
          p = list(fit_ctf,
                   fit_microTensor$unweighted, 
                   fit_microTensor) %>% 
            purrr::map_dbl(
              ~ summary(
                lm(.x$s[, i_axis] ~ rep(c(0, 1), each = n_s / 2))
              )$coef[2, 4]
            ),
          method = c("CTF", "microTensor (uw)", "microTensor"),
          axis = i_axis)
      }) 
    
    tb_pvals <- rbind(tb_pvals_pca, tb_pvals) %>% 
      dplyr::mutate(i_job = ii_job)
    
    l_tb_pvals <- c(l_tb_pvals, list(tb_pvals))
  }
  tb_pvals <- l_tb_pvals %>% 
    purrr::reduce(rbind)
  save(tb_pvals, file = paste0(dir_output, "/pvals_", i_job, ".RData"))
  return(NULL)
}
```

```{r submit jobs}
batchtools::clearRegistry()
tb_ids <- batchtools::batchMap(one_job, 
                               i_job = seq(1, n_job / n_job_each))
batchtools::batchExport(list("dir_output" = dir_output,
                             "n_job_each" = n_job_each,
                             "feature_full" = feature_full,
                             "feature_up" = feature_up,
                             "feature_down" = feature_down))
ncpus <- 1
walltime <- 3600
batchtools::submitJobs(ids = seq(1, 10),
                       resources =  list(ncpus = ncpus,
                                         walltime = walltime))
batchtools::submitJobs(ids = seq(11, n_job / n_job_each),
                       resources =  list(ncpus = ncpus,
                                         partition = partition,
                                         walltime = walltime))
```

```{r summarize results}
library(ggplot2)

load(paste0(dir_output, "/tb_job.RData"))
l_tb_results <- list()
for(i_job in seq(1, nrow(tb_job) / n_job_each)) {
  load(paste0(dir_output, "/pvals_", i_job, ".RData"))
  l_tb_results <- c(l_tb_results, 
                    list(tb_pvals))
}
tb_results <- l_tb_results %>% 
  purrr::reduce(rbind)
tb_results <- tb_results %>% 
  dplyr::left_join(tb_job, by = "i_job")

tb_summary <- tb_results %>%
  dplyr::group_by(strength_disease, 
                  strength_subject, method, n_s, n_t, median_depth, rep) %>% 
  dplyr::summarize(sig_one = any(p[seq(1, 2)] < 0.05 / 2)) %>% 
  dplyr::group_by(strength_disease, strength_subject, method, n_s, n_t, median_depth) %>% 
  dplyr::summarize(power = mean(sig_one),
                   power_sd = sd(sig_one) / sqrt(dplyr::n())) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(Depth = 
                  dplyr::case_when(
                    median_depth == 1e4 ~ "Median 10,000 Reads",
                    TRUE ~ "Median 100,000 Reads"
                  ) %>% 
                  factor(levels = c("Median 10,000 Reads", "Median 100,000 Reads"))) %>% 
  dplyr::arrange(Depth) %>% 
  dplyr::mutate(level_plot_label = 
                  c("A", "B")[as.integer(Depth)] %>% 
                  paste0(") ", Depth) %>% 
                  forcats::as_factor())

colors <- c("PCA" = "black", 
            smar::gg_color_hue(
              values = c("CTF",
                         "microTensor (uw)",
                         "microTensor")))

p <- tb_summary %>% 
  dplyr::mutate(method = factor(method, 
                                levels = names(colors))) %>% 
  dplyr::filter(n_s == 100) %>% 
  ggplot(aes(x = strength_disease,
             y = power,
             color = method,
             shape = method)) +
  geom_point(position = position_dodge(width = 0.1), size = 2) +
  geom_errorbar(aes(ymax = power + power_sd,
                    ymin = power - power_sd),
                position = position_dodge(width = 0.1)) +
  facet_wrap(~ level_plot_label, ncol = 2) +
  scale_color_manual(values = colors, name = "Method") +
  scale_shape_manual(values = c("PCA" = 18,
                                "CTF" = 19,
                                "microTensor (uw)" = 17,
                                "microTensor" = 15), 
                     name = "Method") +
  theme_bw() +
  theme(strip.background = element_blank(),
        legend.position = c(0, 1),
        legend.justification = c(0, 1),
        legend.background = element_blank()) +
  geom_hline(yintercept = 0.05, linetype = "dashed") +
  xlab("Disease effect size") +
  ylab("Power")
ggsave(filename = "figures/figure5_real_world_sim/figure5.pdf",
       p,
       width = 7, height = 4)

p2 <- tb_summary %>% 
  dplyr::mutate(method = factor(method, 
                                levels = names(colors))) %>% 
  dplyr::filter(n_s == 30, 
                strength_disease < 1.5) %>% 
  ggplot(aes(x = strength_disease,
             y = power,
             color = method,
             shape = method)) +
  geom_point(position = position_dodge(width = 0.1), size = 2) +
  geom_errorbar(aes(ymax = power + power_sd,
                    ymin = power - power_sd),
                position = position_dodge(width = 0.1)) +
  facet_wrap(~ level_plot_label, ncol = 2) +
  scale_color_manual(values = colors, name = "Method") +
  scale_shape_manual(values = c("PCA" = 18,
                                "CTF" = 19,
                                "microTensor (uw)" = 17,
                                "microTensor" = 15), 
                     name = "Method") +
  theme_bw() +
  theme(strip.background = element_blank(),
        legend.position = c(0, 1),
        legend.justification = c(0, 1),
        legend.background = element_blank()) +
  geom_hline(yintercept = 0.05, linetype = "dashed") +
  xlab("Disease effect size") +
  ylab("Power")
ggsave(filename = "figures/figure5_real_world_sim/figure5_lessSamples.pdf",
       p2,
       width = 7, height = 4)
```

```{r plot ROC}
tb_results <- tb_results %>% 
  dplyr::filter(strength_disease == 0.75, n_s == 100)

tb_results_summary <- 
  tb_results %>% 
  dplyr::group_by(method, i_job, median_depth) %>% 
  dplyr::group_modify(~ {
    i_tb <- .x
    seq(0, 1, by = 0.05) %>% 
      purrr::map_df(function(cutoff) {
        tibble::tibble(
          cutoff = cutoff,
          sig_one = any(i_tb$p < cutoff / 2)
        )
      })
  })

tb_summary <- tb_results_summary %>% 
  dplyr::left_join(tb_results, by = c("method", "i_job", "median_depth")) %>% 
  dplyr::group_by(strength_disease, method, cutoff, median_depth) %>% 
  dplyr::summarize(power = mean(sig_one),
                   power_sd = sd(sig_one) / sqrt(dplyr::n())) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(Depth = 
                  dplyr::case_when(
                    median_depth == 1e4 ~ "Median 10,000 Reads",
                    TRUE ~ "Median 100,000 Reads"
                  ) %>% 
                  factor(levels = c("Median 10,000 Reads", "Median 100,000 Reads")))

p_ROC <- tb_summary %>% 
  ggplot(aes(x = cutoff, y = power, color = method, shape = method)) +
  geom_point(size = 2) +
  geom_line() +
  scale_color_manual(values = colors, name = "Method") +
  scale_shape_manual(values = c("PCA" = 18,
                                "CTF" = 19,
                                "microTensor (uw)" = 17,
                                "microTensor" = 15), 
                     name = "Method") +
  theme_bw() +
  xlab("Type I Error") + ylab("Power") +
  facet_grid(.~Depth) + 
  theme(legend.position = c(1, 0),
        legend.justification = c(1, 0),
        legend.background = element_blank()) 
ggsave(p_ROC, filename = "figures/figure5_real_world_sim/fig5_ROC.pdf",
       width = 6, height = 3.5)
```
